#137 [remove] 사용하지 않는 api 삭제
