+++ src/main/java/hous/server/controller/user/UserController.java
@@ -5,7 +5,6 @@ import hous.server.common.dto.SuccessResponse;
 import hous.server.config.interceptor.Auth;
 import hous.server.config.resolver.UserId;
 import hous.server.service.user.UserService;
-import hous.server.service.user.dto.request.SetOnboardingInfoRequestDto;
 import hous.server.service.user.dto.request.UpdatePushSettingRequestDto;
 import hous.server.service.user.dto.request.UpdateTestScoreRequestDto;
 import hous.server.service.user.dto.request.UpdateUserInfoRequestDto;
@@ -25,35 +24,6 @@ import javax.validation.Valid;
 public class UserController {
 
     private final UserService userService;
-
-    @ApiOperation(
-            value = "[인증] 온보딩 페이지 - 나의 온보딩 정보를 설정합니다.",
-            notes = "닉네임, 생년월일을 설정합니다. 성공시 status code = 204, 빈 response body를 보냅니다."
-    )
-    @ApiResponses(value = {
-            @ApiResponse(code = 204, message = ""),
-            @ApiResponse(
-                    code = 400,
-                    message = "1. 닉네임을 입력해주세요. (nickname)\n"
-                            + "2. 닉네임은 최대 5글자까지 가능합니다. (nickname)\n"
-                            + "3. 생년월일을 입력해주세요. (birthday)\n"
-                            + "4. 생년월일을 공개 여부를 체크해주세요. (isPublic)",
-                    response = ErrorResponse.class),
-            @ApiResponse(code = 401, message = "토큰이 만료되었습니다. 다시 로그인 해주세요.", response = ErrorResponse.class),
-            @ApiResponse(code = 404,
-                    message = "1. 탈퇴했거나 존재하지 않는 유저입니다. \n"
-                            + "2. 같은 방에 참가하고 있지 않습니다.",
-                    response = ErrorResponse.class),
-            @ApiResponse(code = 500, message = "예상치 못한 서버 에러가 발생하였습니다.", response = ErrorResponse.class)
-    })
-    @Auth
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    @PostMapping("/user/onboarding")
-    public ResponseEntity<String> setOnboardingInfo(
-            @Valid @RequestBody SetOnboardingInfoRequestDto request, @ApiIgnore @UserId Long userId) {
-        userService.setOnboardingInfo(request, userId);
-        return SuccessResponse.NO_CONTENT;
-    }
 
     @ApiOperation(
             value = "[인증] 마이 페이지(Profile 뷰) - 나의 프로필 정보를 수정합니다.",
@@ -84,8 +54,7 @@ public class UserController {
     @Auth
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @PutMapping("/user")
-    public ResponseEntity<String> updateUserInfo(
-            @Valid @RequestBody UpdateUserInfoRequestDto request, @ApiIgnore @UserId Long userId) {
+    public ResponseEntity<String> updateUserInfo(@Valid @RequestBody UpdateUserInfoRequestDto request, @ApiIgnore @UserId Long userId) {
         userService.updateUserInfo(request, userId);
         return SuccessResponse.NO_CONTENT;
     }
